requirements.txt
streamlit==1.32.0
Pillow==10.2.0
google-generativeai==0.3.2
streamlit-drawable-canvas==0.9.3